+++ src/main/java/ru/netology/domain/Book.java
@@ -44,5 +44,3 @@ public class Book extends Product {
                 '}';
     }
 }
-//
-//
