+++ src/main/java/ru/netology/domain/Smartphone.java
@@ -44,5 +44,4 @@ public class Smartphone extends Product {
                 '}';
     }
 }
-//
-//
+
